### 2024-03-24 #6534 @editorialbot
[PRE REVIEW]: lightcurver: A Python pipeline for precise photometry of multiple-epoch wide-field images
**Submitting author:** @duxfrederic (<a href="http://orcid.org/0000-0003-3358-4834">Frédéric Dux</a>)
**Repository:** https://github.com/duxfrederic/lightcurver
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.3
