### 2024-03-25 be3d7be
added possibility to include backgrounds when doing star photometry
