### 2024-03-30 731eb3f
added utility function to bin light curves in later stages
