### 2024-03-31 30a186b
utility function to convert flux and flux errors to mags
