### 2024-04-01 45a94a9
ignoring absence of scatter in utility function for flux to mag
