### 2024-04-01 b5fcee8
convention change in pandas dataframe of lightcurves
