### 2024-04-01 6de407c
convention change in pandas dataframe of lightcurves
