### 2024-04-09 880faab
passing fwhm estimate to psf routine
