### 2024-04-19 3a85605
filter on normalization
