### 2024-05-01 9f13f4e
added approx zeropoint calibration on more bands thanks to gaia
