### 2024-05-23 23de9cc
consistent rescaled coordinates for psf building
