### 2024-05-23 35dc5b7
photometry of stars: using distorted psf if required. tests: includes pipeline with distortion
