+++ tests/test_processes/test_cutout_making.py
@@ -29,7 +29,7 @@ class TestExtractStamp(unittest.TestCase):
 
     def test_extract_stamp(self):
         # check the function can be called
-        cutout, noisemap, wcs_header_string = extract_stamp(
+        cutout, noisemap, wcs_header_string, position = extract_stamp(
             self.data, self.header, self.exptime, self.sky_coord, self.cutout_size,
             background_rms_electron_per_second=0.5)
 
