### 2024-05-23 21553bd
roi files prep: psf distortion if requested
