+++ lightcurver/processes/roi_deconv_file_preparation.py
@@ -149,7 +149,20 @@ def prepare_roi_deconv_file():
             noisemap.append(h5f[f"{frame['image_relpath']}/noisemap/ROI"][...] / coefficient)
             mask.append(h5f[f"{frame['image_relpath']}/cosmicsmask/ROI"][...])
             psf_ref = frame['psf_ref']
-            psf.append(h5f[f"{frame['image_relpath']}/{psf_ref}/narrow_psf"][...])
+            narrow_psf = h5f[f"{frame['image_relpath']}/{psf_ref}/narrow_psf"][...]
+            # if distortion of the psf ...
+            if user_config['field_distortion']:
+                kwargs_distortion = {}
+                for key, param in h5f[f"{frame['image_relpath']}/{psf_ref}/distortion"]:
+                    kwargs_distortion[key] = param
+                # 2. collect the position of the star in the frame
+                position = h5f[f"{frame['image_relpath']}/image_pixel_coordinates/{star['gaia_id']}"][...]
+                frame_shape = h5f[f"{frame['image_relpath']}/frame_shape"]
+                position = rescale_image_coordinates(xy_coordinates_array=position, image_shape=frame_shape)
+                # 3. get the psf at this position
+                narrow_psf = apply_distortion(narrow_psf=narrow_psf, kwargs_distortion=kwargs_distortion,
+                                              star_xy_coordinates=position)
+            psf.append(narrow_psf)
             subsampling_factors.append(h5f[f"{frame['image_relpath']}/{psf_ref}/subsampling_factor"][...])
             seeing.append(frame['seeing_arcseconds'])
             pixel_scale.append(frame['pixel_scale'])
