### 2024-05-24 ccf6d8f
roi files prep: psf distortion if requested
