+++ lightcurver/processes/roi_deconv_file_preparation.py
@@ -2,10 +2,13 @@ import h5py
 import numpy as np
 import logging
 
+from starred.psf.psf import apply_distortion
+
 from ..structure.database import get_pandas, execute_sqlite_query
 from ..utilities.footprint import get_combined_footprint_hash
 from ..utilities.chi2_selector import get_chi2_bounds
 from ..structure.user_config import get_user_config
+from ..utilities.image_coordinates import rescale_image_coordinates
 
 
 def get_frames_for_roi(combined_footprint_hash, psf_fit_chi2_min, psf_fit_chi2_max,
@@ -153,10 +156,11 @@ def prepare_roi_deconv_file():
             # if distortion of the psf ...
             if user_config['field_distortion']:
                 kwargs_distortion = {}
-                for key, param in h5f[f"{frame['image_relpath']}/{psf_ref}/distortion"]:
-                    kwargs_distortion[key] = param
+                distortion_group = h5f[f"{frame['image_relpath']}/{psf_ref}/distortion"]
+                for key in distortion_group.keys():
+                    kwargs_distortion[key] = distortion_group[key][...]
                 # 2. collect the position of the star in the frame
-                position = h5f[f"{frame['image_relpath']}/image_pixel_coordinates/{star['gaia_id']}"][...]
+                position = h5f[f"{frame['image_relpath']}/image_pixel_coordinates/ROI"][...]
                 frame_shape = h5f[f"{frame['image_relpath']}/frame_shape"]
                 position = rescale_image_coordinates(xy_coordinates_array=position, image_shape=frame_shape)
                 # 3. get the psf at this position
