### 2024-05-24 e0d570b
switched to southern image: indicate in paper
