### 2024-05-27 0c1bc4e
working tests: adapting WCS
