### 2024-05-27 3f09594
logger was not called correctly
