+++ lightcurver/utilities/footprint.py
@@ -1,0 +1,219 @@
+import numpy as np
+from shapely.geometry import Polygon, mapping
+from functools import reduce
+import json
+from astropy.io import fits
+from astropy.wcs import WCS
+
+from ..structure.database import execute_sqlite_query, get_pandas
+from ..structure.user_config import get_user_config
+
+
+def get_combined_footprint_hash(user_config, frames_id_list):
+    """
+    Calculates the hash of the combined footprint of the frames whose id is in frames_id_list.
+    Args:
+        user_config: dictionary obtained from structure.config.get_user_config
+        frames_id_list: list of integers, which frames are we using.
+
+    Returns:
+        frames_hash, integer
+
+    """
+    if user_config['star_selection_strategy'] != 'ROI_disk':
+        # then it depends on the frames we're considering.
+        frames_hash = get_frames_hash(frames_id_list)
+    else:
+        # if ROI_disk, it does not depend on the frames: unique region defined by its radius.
+        return hash(user_config['ROI_disk_radius_arcseconds'])
+
+
+def calc_common_and_total_footprint(list_of_footprints):
+    """
+    Calculate the common (intersection) and largest (union) footprints from a list of numpy arrays: products
+    of astropy.wcs.WCS.calc_footprint.
+    Then determines both the intersection and union of these
+    footprints using Shapely Polygons. The intersection represents the area common to all frames,
+    while the union covers the total area spanned by any of the frames.
+
+    Parameters:
+    - list_of_fits_paths: A list of paths to FITS files.
+
+    Returns:
+    - wcs_objects: A list of WCS objects corresponding to the FITS files.
+    - common_footprint: A Shapely Polygon representing the intersected footprint common to all WCS footprints.
+    - largest_footprint: A Shapely Polygon representing the union of all WCS footprints.
+    """
+    wcs_footprints = list_of_footprints
+
+    polygons = [Polygon(footprint) for footprint in wcs_footprints]
+
+    common_footprint = reduce(lambda x, y: x.intersection(y), polygons)
+    largest_footprint = reduce(lambda x, y: x.union(y), polygons)
+
+    return (common_footprint.simplify(tolerance=0.001, preserve_topology=True),
+            largest_footprint.simplify(tolerance=0.001, preserve_topology=True))
+
+
+def database_insert_single_footprint(frame_id, footprint_array):
+    polygon_list = footprint_array.tolist()
+    polygon_str = json.dumps(polygon_list)
+
+    execute_sqlite_query(query="INSERT OR REPLACE INTO footprints (frame_id, polygon) VALUES (?, ?)",
+                         params=(frame_id, polygon_str),
+                         is_select=False)
+
+
+def database_get_footprint(frame_id):
+    result = execute_sqlite_query(query="SELECT polygon FROM footprints WHERE frame_id = ?",
+                                  params=(frame_id,),
+                                  is_select=True)[0]
+
+    polygon_list = json.loads(result)
+    footprint_polygon = np.array(polygon_list)
+
+    return footprint_polygon
+
+
+def get_frames_hash(frames_ids):
+    """
+    when calculating footprints, we need a way to identify which footprint was calculated from which frames.
+    I don't want to deal with the relational many-to-many situation that will arise in a relational database,
+    so let's calculate a hash of the integers of the frames that were used to calculate a footprint.
+    then to check for a footprint, which can just query the hash.
+    Args:
+        frames_ids: list of integers, frames.id in the database
+
+    Returns:
+        a text hash
+
+    """
[…]
