+++ paper.bib
@@ -1,0 +1,276 @@
+@ARTICLE{starred,
+       author = {{Michalewicz}, Kevin and {Millon}, Martin and {Dux}, Fr{\'e}d{\'e}ric and {Courbin}, Fr{\'e}d{\'e}ric},
+        title = "{STARRED: a two-channel deconvolution method with Starlet regularization}",
+      journal = {The Journal of Open Source Software},
+     keywords = {Python, PSF, Jupyter Notebook, astronomy, deconvolution},
+         year = 2023,
+        month = may,
+       volume = {8},
+       number = {85},
+          eid = {5340},
+        pages = {5340},
+          doi = {10.21105/joss.05340},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023JOSS....8.5340M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{starredscience,
+       author = {{Millon}, Martin and {Michalewicz}, Kevin and {Dux}, Fr{\'e}d{\'e}ric and {Courbin}, Fr{\'e}d{\'e}ric and {Marshall}, Philip J.},
+        title = "{Image deconvolution and PSF reconstruction with STARRED: a wavelet-based two-channel method optimized for light curve extraction}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
+         year = 2024,
+        month = feb,
+          eid = {arXiv:2402.08725},
+        pages = {arXiv:2402.08725},
+          doi = {10.48550/arXiv.2402.08725},
+archivePrefix = {arXiv},
+       eprint = {2402.08725},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024arXiv240208725M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
+@article{gaia,
+    author = {{Gaia Collaboration}},
+    title = "{The Gaia mission}",
+    journal = {Astronomy and Astrophysics},
+    archivePrefix = "arXiv",
+    eprint = {1609.04153},
+    primaryClass = "astro-ph.IM",
+    keywords = {space vehicles: instruments, Galaxy: structure, astrometry, parallaxes, proper motions, telescopes},
+    year = 2016,
+    month = nov,
+    volume = 595,
+    doi = {10.1051/0004-6361/201629272},
+    url = {http://adsabs.harvard.edu/abs/2016A%26A...595A...1G},
+}
+
+@article{LSST,
+	doi = {10.3847/25c2cfeb.d8909f28},
+	author = {Vera C. Rubin Observatory LSST Solar System Science Collaboration, and Jones, R. Lynne and Bannister, Michelle T. and Bolin, Bryce T. and Chandler, Colin Orion and Chesley, Steven R. and Eggl, Siegfried and Greenstreet, Sarah and Holt, Timothy R. and Hsieh, Henry H. and Ivezic, Zeljko and Juric, Mario and Kelley, Michael S. P. and Knight, Matthew M. and Malhotra, Renu and Oldroyd, William J. and Sarid, Gal and Schwamb, Megan E. and Snodgrass, Colin and Solontoi, Michael and Trilling, David E.},
+	journal = {Bulletin of the AAS},
+	number = {4},
+	year = {2021},
+	month = {mar 18},
+	note = {https://baas.aas.org/pub/2021n4i236},
+	publisher = {},
+	title = {The {Scientific} {Impact} of the {Vera} {C}. {Rubin} {Observatory}\textquoteright{}s {Legacy} {Survey} of {Space} and {Time} ({LSST}) for {Solar} {System} {Science}},
+	volume = {53},
+}
+
[…]

[…]
