### 2024-05-27 #6775 @duxfrederic
Hi @ivastar
Thanks for checking, ended up completely synchronizing the paper branch with what's in the main one at the moment.
