### 2024-06-17 #6775 @duxfrederic
Hey @Onoddil,
thanks for starting your work on this!

Regarding the Reproducibility check item, [this](https://drive.google.com/file/d/1788snqvZS676Upzc9L9pfNzf2WQhnF07/view?usp=sharing) is the directory structure produced by the pipeline to make the example result of shown in the paper.
It includes:
- all the cutouts made from the calibrated images (in `regions.h5`)
- the PSFs produced from the calibrated images  (also in `regions.h5`)
- the sqlite3 database
- the cutouts + PSFs of the ROI in `prepared_cutouts/cutouts_195_PSJ0030-1525.h5`, together with a jupyter notebook producing the light curves and deconvolved image from these cutouts.
- the diagnostic plots, logs
- the config file

They do not include the calibrated images themselves, which are quite voluminous (39 GB). If you feel like you need them to check the item, I'll find a way to send them over!
However, from this, all the steps can be tested except from those involving the calibrated images themselves (importation, plate solving, cutout making)
