### 2024-06-17 duxfrederic/lightcurver#28 @Onoddil
Cannot run software installed via `pip install` due to dependency clash
At present -- admittedly _possibly_ a one-off issue caused by `numpy`'s V2.0 release -- I cannot install `lightcurver` as described in the documentation due to inflexible `pip` constraints.

This boils down to `sep`, amongst other released packages, likely targetting `numpy` on its `1.26` release and `pip` picking up the `2.0.0` tag, and hence the `pytest` error on a git-cloned repo:
```
lightcurver/processes/psf_modelling.py:4: in <module>
    import sep
sep.pyx:1: in init sep
    ???
E   ValueError: numpy.dtype size changed, may indicate binary incompatibility. Expected 96 from C header, got 88 from PyObject
```
or, after `pip install lightcurver` from a fresh `conda` environment:
```
python -c "import sep"
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "sep.pyx", line 1, in init sep
ValueError: numpy.dtype size changed, may indicate binary incompatibility. Expected 96 from C header, got 88 from PyObject
```

To get around this, I had to ignore the limited installation instructions, and manually `conda install` the "big" dependencies, to make sure I picked up a compatible set of releases for things like `numpy`, `astropy`, `sep`, etc., before running `pip install lightcurver` (or `pip install .` in the case of a `git clone`). Once I had made sure of this large-scale compatibility, then the codebase was installed successfully and the tests ran without issue.

Again, I _suspect_ this is just an unfortunate step-change in dependency with the big `numpy` release, but the `numpy.dtype size changed` issue does surface from time to time, and you may want to expand the documentation to explicitly just list the required dependencies (i.e., those from `pyproject.toml` that I looked at to generate the `conda install` list I used to try to guarantee inter-dependency) in the documentation, and -- in addition to providing users the "will usually work" `pip install lightcurver` option -- give them a slower, but more likely to work in these cases, manual-stage `conda install` setup for these larger dependencies. Of the smaller dependencies you may or may not need to manually `conda install` them to ensure robust inter-package dependency resolution, or may need to `pip install` them separately (or within the final `pip install lightcurver` step) depending on whether they are released through conda channels.

(see [JOSS review](https://github.com/openjournals/joss-reviews/issues/6775))
