### 2024-06-17 duxfrederic/lightcurver#28 @duxfrederic
Wow, amazing insight, thanks!

This indeed fails on my side:
```
conda create -n lightcurver_env python=3.12
conda activate lightcurver_env
pip install lightcurver
python -c "import sep"
```
and it is fixed by
```
pip install "numpy<2.00"
```

So, the fix:

- explicitly set `numpy<2.00` in the `pyproject.toml` until `sep` becomes compatible with `numpy 2.0`,
- as per your recommendation, expand the documentation for manual installation
