### 2024-06-17 e3300cb
added list of dependencies to docs
