### 2024-06-17 4478bdb
removed starred dependencies from pyproject.toml (fixed starred install of its own deps); and enforces numpy < 2.00
