+++ pyproject.toml
@@ -14,7 +14,7 @@ dependencies = [
     "pyyaml",
     "matplotlib",
     "astropy",
-    "numpy",
+    "numpy < 2.0.0",
     "sep",
     "scipy",
     "ephem",
@@ -25,16 +25,9 @@ dependencies = [
     "astroscrappy",
     "starred-astro >= 1.4.2", 
     "pytest",
-    "dill",
-    "jax",
-    "jaxlib",
-    "jaxopt",
     "matplotlib",
     "numpy",
     "scipy",
-    "optax",
-    "tqdm",
-    "emcee",
     "h5py",
     "widefield_plate_solver", 
     "photutils",
