### 2024-06-17 065f785
docs: more comments on dependencies
