+++ docs/mkdocs/contents/installation.md
@@ -36,22 +36,22 @@ Or for the `git` version (includes some minimal test data):
 Should the above fail, there might be a dependency problem requiring the manual handling of the different packages. 
 Here is the list of dependencies that need be installed:
 
-1. `numpy < 2.00`
+1. `numpy < 2.00` - as of June 2024, `sep` is not compatible with `numpy >= 2.00`
 2. `scipy`
 3. `matplotlib`
 4. `pandas`
 5. `astropy`
-6. `astroquery`
-7. `h5py`
-8. `photutils`
-9. `astroalign`
-10. `shapely`
-11. `ephem (pyephem)`
-12. `pytest`
-13. `sep`
-14. `astroscrappy`
-15. `pyyaml`
-16. `starred-astro` - assume the latest version
+6. `astroquery` - for querying Gaia and VizieR
+7. `h5py` - for storing cutouts and PSF models
+8. `photutils` - for aperture photometry used as initial guess
+9. `astroalign` - for finding transformations between frames
+10. `shapely` - for computing footprints of frames
+11. `ephem` - for calculating airmass, moon distance, etc.
+12. `pytest` - for executing the automatic tests
+13. `sep` - for source and background extraction
+14. `astroscrappy` - for cleaning the cosmics
+15. `pyyaml` - for reading the config file
+16. `starred-astro` - assume the latest version, will install its own dependencies.
 17. `widefield_plate_solver` - an astrometry.net wrapper
 
 
