### 2024-06-17 duxfrederic/lightcurver#28 @duxfrederic
@Onoddil the PR above should have addressed your comments, let me know if not

(I will wait until the end of the review to update the pypi repository, though)
