### 2024-06-20 duxfrederic/lightcurver#28 @Onoddil
Yes, that fixes it.

Of course, the real trick is remembering to go back and update the dependencies once `sep` updates upstream, but providing the list of dependencies on the website should give users enough to go on to solve the problem themselves if it comes up again :)

You can't really do much more than that, otherwise it quickly becomes a full time job just pushing minor revisions for package dependencies I guess!
