### 2024-06-20 #6775 @Onoddil
Figuring it makes more sense to leave here, being JOSS paper-related rather than codebase-related (and hence split out into an issue over on the main repo), but just a few missing references that should ideally be included, if you have the room:

- astropy: you included the recent citation, but if possible you could include the other two citations (Astropy Collaboration et al. 2013, A&A, 558, 33; Astropy Collaboration et al. 2018, ApJ, 156, 123) -- see https://www.astropy.org/acknowledging.html for BibTeX.
- matplotlib: requests a citation for Hunter 2007 (Computing in Science & Engineering, 9, 90) -- https://matplotlib.org/stable/project/citing.html for .bib
- numpy: Harris et al. 2020 (Nature, 585, 357) -- https://numpy.org/citing-numpy/ for .bib
- scipy: Virtanen et al. 2020 (Nature Methods, 17, 261) -- https://scipy.org/citing-scipy/ for .bib
- pandas: The pandas development team 2020 (doi.org/10.5281/zenodo.3509134) and Data structures for statistical computing in python, McKinney, Proceedings of the 9th Python in Science Conference, Volume 445, 2010 -- see https://pandas.pydata.org/about/citing.html
- photutils: Requests "This research made use of Photutils, an Astropy package for detection and photometry of astronomical sources (Bradley et al. <YEAR>)." as text; see https://zenodo.org/records/10967176 for citation and .bib
- astroalign: "Astroalign: A Python module for astronomical image registration." Beroiz, M., Cabral, J. B., & Sanchez, B. Astronomy and Computing, Volume 32, July 2020, 100384 (https://astroalign.quatrope.org/en/latest/#citation for .bib)
- starred-astro: Michalewicz et al., (2023). STARRED: a two-channel deconvolution method with Starlet regularization. Journal of Open Source Software, 8(85), 5340 -- https://joss.theoj.org/papers/10.21105/joss.05340 for bibtex

These can presumably just all go in a very dense sentence/paragraph of "this software makes use of X \cite{}, Y \cite{}, Z \cite{}, ..." in the appropriate place.
