### 2024-06-20 c073f80
adding references to packages lightcurver relies on
