### 2024-06-21 #6775 @Onoddil
You've accidentally cited the 2022 astropy paper twice instead of the 2013 one, but apart from that this all LGTM :)
