### 2024-06-21 46e4e16
astropy citation bugged? trying different naming of bib keys
