### 2024-07-09 a3e97bc
extra database structure for external catalogs of absolute photometry
