### 2024-07-09 5e6d869
comment in gaia file to explain its difference with future files of other surveys
