### 2024-07-09 2ea89e8
added structure for pan-starrs, not integrated in pipeline yet, missing tests as well
