### 2024-07-09 692fdbc
renamed database approximate to absolute zeropoints, better naming given additional surveys
