### 2024-07-09 cdddff2
wrote factored out version of zero point calculation
