### 2024-07-09 e914377
made it clearer what is happening when using gaia or panstarrs, kept refactoring. missing: update workflow manager and tutorial
