### 2024-07-09 5b1b689
corrected: query star magnitudes before trying to calculate zeropoints
