+++ lightcurver/processes/roi_deconv_file_preparation.py
@@ -93,7 +93,8 @@ def fetch_and_adjust_zeropoints(combined_footprint_hash):
     WHERE
         az.combined_footprint_hash = ?
     """
-    zeropoints_data = execute_sqlite_query(zeropoint_query, (combined_footprint_hash,), is_select=True, use_pandas=True)
+    zeropoints_data = execute_sqlite_query(zeropoint_query, (combined_footprint_hash,),
+                                           is_select=True, use_pandas=True)
 
     if zeropoints_data.empty:
         return None, None
