### 2024-07-09 de1e65b
corrected insert queries and added filtering of panstarrs
