### 2024-07-10 8c88db6
moving example config inside package directory for setuptools inclusion
