+++ pyproject.toml
@@ -31,7 +31,8 @@ dependencies = [
     "h5py",
     "widefield_plate_solver", 
     "photutils",
-    "astroalign"
+    "astroalign",
+    "ruamel.yaml"
 ]
 
 [project.urls]
@@ -46,4 +47,5 @@ test = [
 [tool.setuptools]
 packages.find = {where = ["."], include = ["lightcurver*"]}
 package-dir = {"" = "."}
-package-data = {"lightcurver" = ["pipeline/*.yaml"]}
+package-data = {"lightcurver" = ["pipeline/*.yaml", "pipeline/example_config_file/config.yaml"]}
+
