### 2024-07-10 fcc26af
adapted link to config in tutorial
