### 2024-07-10 3fd94e7
added 2 scripts interfaces
