### 2024-07-10 6e97e18
scripts entry points in pyproject
