### 2024-07-10 bfb7eb8
furthter removed filter from database and tutorial
