### 2024-07-10 f4e265f
further removed filter mentions; corrected new path to template config in tests
