### 2024-07-15 #6775 @ivastar
@duxfrederic I've just done that. I think they were not seeing the GitHub notifications, but should be done soon.
