### 2024-07-17 ece34ff
bug fix: redo cutout coordinate storage of star was deleting roi and not star
