### 2024-07-17 28d6fb0
safety: str(gaia_id) before insertion into sqlite
