### 2024-07-31 4ca6168
edit to run on DECAM images
