+++ lightcurver/processes/frame_importation.py
@@ -24,31 +24,31 @@ def process_new_frame(fits_file, user_config):
     """
     trim_vertical = user_config.get('trim_vertical', 0)
     trim_horizontal = user_config.get('trim_horizontal', 0)
+    hdu_index = user_config.get('hdu_index', 0)
+    hdu_index_header = user_config.get('hdu_index_header', 0)
     copied_image_relpath = Path('frames') / f"{fits_file.stem}.fits"
     logger = logging.getLogger("lightcurver.importation")
 
     try:
         with fits.open(str(fits_file), mode='readonly', ignore_missing_end=False, memmap=True) as hdu:
             logger.info(f'  Importing {fits_file}.')
-            hdu_index = 1 if len(hdu) > 1 else 0
             data_shape = hdu[hdu_index].data.shape
             cutout_data = hdu[hdu_index].section[
                           trim_vertical:data_shape[0] - trim_vertical,
                           trim_horizontal:data_shape[1] - trim_horizontal
                           ]
-            header = hdu[hdu_index].header
+            header = hdu[hdu_index_header].header
             logger.info(f"Read fits file at {fits_file}")
     except ValueError:
         # then we have some memmap problems, if  bzero, bscale, or blank are in header.
         with fits.open(str(fits_file), mode='readonly') as hdu:
             logger.info(f'  Importing {fits_file}.')
-            hdu_index = 1 if len(hdu) > 1 else 0
             data_shape = hdu[hdu_index].data.shape
             cutout_data = hdu[hdu_index].data[
                           trim_vertical:data_shape[0] - trim_vertical,
                           trim_horizontal:data_shape[1] - trim_horizontal
                           ]
-            header = hdu[hdu_index].header
+            header = hdu[hdu_index_header].header
             logger.warning(f"Read fits file at {fits_file}, could not use memmap!!")
     cutout_data = cutout_data.astype(float)
     wcs = WCS(header)
