### 2024-07-31 b02f689
edit to run on DECAM images
