### 2024-08-12 #6775 @duxfrederic
Hi @ivastar

thanks a bunch for contacting our second reviewer last month, have you had news from them? :)
