### 2024-08-21 #6775 @duxfrederic
No worries at all, I know JOSS is run by volunteers with other things going on!

Regarding suggestions, going through the list of reviewers:

- maybe benjaminrose, they do supernova -- although they probably have their own methods for photometry, they might find this code base somewhat interesting.
- Gabriel-p : they do more advanced things than photometry, but I noticed they do have an interest in photometry as well
- tbmiller-astro seem to be working with images in their research, maybe they'd be interested
- samvaughan seem to work with a lot of astronomical data as well, again maybe an interest!

thanks!
