### 2024-09-06 #6775 @benjaminrose
I am getting an error from multiprocessing. I have tried python 3.12, 3.10, installing with pip and installing from source. I hit the same error on `initialize_database()` and `WorkflowManager().run()`. The full traceback is below. I am not sure what is happening or where exactly the problem is, but currently I am unable to test the functionality of lightcurver.

  <summary>Traceback</summary>

```
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/spawn.py", line 116, in spawn_main
    exitcode = _main(fd, parent_sentinel)
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/spawn.py", line 125, in _main
    prepare(preparation_data)
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/spawn.py", line 236, in prepare
    _fixup_main_from_path(data['init_main_from_path'])
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/spawn.py", line 287, in _fixup_main_from_path
    main_content = runpy.run_path(main_path,
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/runpy.py", line 289, in run_path
    return _run_module_code(code, init_globals, run_name,
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/runpy.py", line 96, in _run_module_code
    _run_code(code, mod_globals, init_globals,
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/runpy.py", line 86, in _run_code
    exec(code, run_globals)
  File "/Users/ben_rose/Desktop/lightcurver_JOSS_REVIEW/LitghtCurve_JOSS_review.py", line 18, in <module>
    read_convert_skysub_character_catalog()
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/site-packages/lightcurver/pipeline/task_wrappers.py", line 51, in read_convert_skysub_character_catalog
    log_queue = Manager().Queue()
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/context.py", line 57, in Manager
    m.start()
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/managers.py", line 562, in start
    self._process.start()
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/process.py", line 121, in start
    self._popen = self._Popen(self)
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/context.py", line 288, in _Popen
    return Popen(process_obj)
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/popen_spawn_posix.py", line 32, in __init__
    super().__init__(process_obj)
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/popen_fork.py", line 19, in __init__
    self._launch(process_obj)
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/popen_spawn_posix.py", line 42, in _launch
    prep_data = spawn.get_preparation_data(process_obj._name)
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/spawn.py", line 154, in get_preparation_data
    _check_not_importing_main()
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/spawn.py", line 134, in _check_not_importing_main
    raise RuntimeError('''
RuntimeError:
        An attempt has been made to start a new process before the
        current process has finished its bootstrapping phase.

        This probably means that you are not using fork to start your
        child processes and you have forgotten to use the proper idiom
        in the main module:

            if __name__ == '__main__':
                freeze_support()
                ...

        The "freeze_support()" line can be omitted if the program
        is not going to be frozen to produce an executable.
Traceback (most recent call last):
  File "/Users/ben_rose/Desktop/lightcurver_JOSS_REVIEW/LitghtCurve_JOSS_review.py", line 18, in <module>
    read_convert_skysub_character_catalog()
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/site-packages/lightcurver/pipeline/task_wrappers.py", line 51, in read_convert_skysub_character_catalog
    log_queue = Manager().Queue()
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/context.py", line 57, in Manager
    m.start()
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/managers.py", line 566, in start
    self._address = reader.recv()
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/connection.py", line 250, in recv
    buf = self._recv_bytes()
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/connection.py", line 414, in _recv_bytes
    buf = self._recv(4)
  File "/Users/ben_rose/mambaforge/envs/lightcurver_env/lib/python3.10/multiprocessing/connection.py", line 383, in _recv
    raise EOFError
EOFError
```
