### 2024-09-07 #6775 @duxfrederic
Hi @benjaminrose

Many thanks for starting this review so fast, and for trying different environments and python versions.

I just got access to a mac computer and I can reproduce your error! I think the automated tests should run smoothly on your end, but the error is due to the way mac and windows computers spawn processes.
With these, you would need to wrap any part of code that runs multiprocessing in a `if __name__ == "__main__"` block.

I'll update the documentation to reflect this; in the mean time you have 3 possibilities:

1. Run the same code in a jupyter notebook rather than in a script.
2. Wrap the function calls in a `if __name__ == "__main__"` block:
```python
import os
# replace with your actual path:
os.environ['LIGHTCURVER_CONFIG'] = "/scratch/lightcurver_tutorial/config.yaml"

from lightcurver.structure.user_config import get_user_config
from lightcurver.structure.database import initialize_database
from lightcurver.pipeline.task_wrappers import read_convert_skysub_character_catalog

if __name__ == "__main__":
    initialize_database()
    read_convert_skysub_character_catalog()
```
or to run everything

```python
import os
os.environ['LIGHTCURVER_CONFIG'] = "/path/to/config.yaml"

from lightcurver.pipeline.workflow_manager import WorkflowManager

if __name__ == "__main__":
    wf_manager = WorkflowManager()
    wf_manager.run()
```
3. Use `lc_run /path/to/your/config.yaml` to run the entire pipeline.

Sorry about this, I should have tested on mac computers more carefully!

Let me know if this fixes things on your side ...
