### 2024-09-21 #6775 @duxfrederic
Hi @benjaminrose

Did the above work? If it is more convenient for you in testing the steps individually, as in the tutorial, I put them together in the scripts attached here:
[lightcurver_scripts_for_review.zip](https://github.com/user-attachments/files/17085807/lightcurver_scripts_for_review.zip)

First:
```bash
export REVIEW_WORKDIR='/tmp/workdir'
export LIGHTCURVER_CONFIG='/tmp/workdir/config.yaml'  # config has to be in workdir in this specific case
```

Then executing the scripts above one after the other worked on the mac I have at hand (M3 Pro, Sonoma 14.5)

But the simpler version is to just run `lc_run config.yaml`.
