### 2024-09-27 94325bf
cosmics masking params in config
