### 2024-09-29 #6775 @benjaminrose
Sorry for not getting back to this as quickly as I should have. I was able to test this on a linux machine and the functionality and performance was as quoted. It looks good to me.
