### 2024-09-30 #6775 @Onoddil
> @benjaminrose and @Onoddil, can I can an "aye" that this paper is accepted?

I assume my previous aye still holds, so yes, aye from me!
