### 2024-09-30 #6775 @benjaminrose
> @benjaminrose and @Onoddil, can I can an "aye" that this paper is accepted?

Yes.
