### 2024-10-02 #6775 @ivastar
@duxfrederic some minor editorial changes. After you've implemented these, please proceed to the checklist below.

- Title: please use proper capitalization
- Line 7: cadenced --> time series
- Line 9: Cepheids needs to be capitalized
- Line 10: remove hyphen in general purpose
- Line 11: "tailored for" --> "tailored to"
- Line 15: "Point Spread Function" --> "point spread function"
- Line 16: "combining PSF-photometry fluxes of several stars in the field of view" --> comparing(?) the fluxes of several stars in the field of view measured via PSF photometry
- Line 19: remove hyphen in "high-resolution"
- Line 40: "zero-point" --> "zeropoint"
- Line 40: "typically a milimag" --> "typically one milimag"
- Line 44: there are 2 periods, remove one
- Line 47: "Firstly" --> "First"
- Line 53-54: "rotations need not be stable across epochs" --> "do not need to be stable across epochs"
- Line 56: which astropy package is used for the photometry? is it astropy or astropy affiliated? If the latter, maybe the package itself has citation instructions.
- Line 62: "spoiling" --> "corrupting"?
- Line 64: "flux it proceeds with" --> "flux to proceed with" or "flux they proceed with"
- Line 65: "subpackages" --> sub-packages
- Line 67: "structure, which contains" --> "structure, containing" (to match the structure of the preceding 2 clauses)
- Line 78: "whose morphology" --> "the morphology of which"
- Line 82: "modeling" --> "modelling"
