+++ paper.bib
@@ -417,7 +417,7 @@ abstract = {A new method for improving the resolution of astronomical images is 
 }
 
 @software{reback2020pandas,
-    author       = {The pandas development team},
+    author       = {development team pandas},
     title        = {pandas-dev/pandas: Pandas},
     month        = feb,
     year         = 2020,
@@ -454,7 +454,7 @@ volume = "32",
 pages = "100384",
 year = "2020",
 issn = "2213-1337",
-doi = "https://doi.org/10.1016/j.ascom.2020.100384",
+doi = "10.1016/j.ascom.2020.100384",
 url = "http://www.sciencedirect.com/science/article/pii/S221313372030038X",
 author = "M. Beroiz and J.B. Cabral and B. Sanchez",
 keywords = "Astronomy, Image registration, Python package",
