+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: '`lightcurver`: A Python pipeline for precise photometry of multiple-epoch wide-field images'
+title: '`lightcurver`: A Python Pipeline for Precise Photometry of Multiple-Epoch Wide-Field Images'
 tags:
   - Python
   - astronomy
@@ -21,14 +21,14 @@ bibliography: paper.bib
 
 # Summary
 
-`lightcurver` is a photometry pipeline for cadenced astronomical imaging data, 
+`lightcurver` is a photometry pipeline for time series astronomical imaging data, 
 designed for the semi-automatic extraction of precise light curves from small, blended targets. 
-Such targets include, but are not limited to, lensed quasars, supernovae, or cepheids in crowded fields. 
-`lightcurver` is not a general-purpose photometry, astrometry, and classification pipeline like `legacypipe` [@legacypipe]. 
-Instead, it is a framework tailored for the precise study of a small region of interest (ROI) in wide-field images, 
+Such targets include, but are not limited to, lensed quasars, supernovae, or Cepheids in crowded fields. 
+`lightcurver` is not a general purpose photometry, astrometry, and classification pipeline like `legacypipe` [@legacypipe]. 
+Instead, it is a framework tailored to the precise study of a small region of interest (ROI) in wide-field images, 
 utilizing stars surrounding the ROI to calibrate the frames.
 
-At its core, `lightcurver` leverages `STARRED` [@starred; @starredscience] to generate state-of-the-art Point Spread Function (PSF) models for each image. 
+At its core, `lightcurver` leverages `STARRED` [@starred; @starredscience] to generate state-of-the-art empirical point spread function (PSF) models for each image. 
 It then determines the relative zeropoints between images by combining the PSF-photometry fluxes of several stars in the field of view. 
 Subsequently, `STARRED` is used again to simultaneously model the calibrated pixels of the ROI across all epochs. 
 This process yields light curves of the point sources and a high-resolution image model of the ROI, cumulating the signal from all epochs.
@@ -51,14 +51,14 @@ requires too much manual intervention to be run on a daily basis.
 On the other hand, `STARRED` is a powerful PSF modelling and deconvolution package, ideal for this task. 
 However, by its nature, it cannot include an infrastructure that makes it convenient to apply to large datasets without manual intervention
 (e.g., visually identifying appropriate stars, extracting cutouts, and all subsequent processing steps leading to a light curve). 
-Particularly, `STARRED` modelling requires a very stable zero-point across modelled epochs, 
+Particularly, `STARRED` modelling requires a very stable zeropoint across modelled epochs, 
 as it emulates the constant components of the ROI as one grid of pixels common to all epochs, 
 which it simultaneously optimizes together with the fluxes of the variables. 
-Achieving such precise relative zero-point calibration (typically a milimag), especially in an automated manner, comes with challenges.
+Achieving such precise relative zeropoint calibration (typically one milimag), especially in an automated manner, comes with challenges.
 
 `lightcurver` addresses this challenge by automatically selecting calibration stars, modelling them, 
 and robustly combining their fluxes to calibrate the zeropoints, 
-making it suitable as a daily running pipeline on a large number of ROIs..
+making it suitable as a daily running pipeline on a large number of ROIs.
 
 
 ![Light curve of a lensed image of a quasar (J0030-1525), extracted once with the existing code base (`COSMOULINE`), 
@@ -70,11 +70,11 @@ HST image: PI Tommaso Treu, proposal GO 15652.](plot/comparison_with_legacy_pipe
 
 `lightcurver` utilizes an SQLite3 database to track data processing stages and relies on SQL queries to manage its workflow, 
 identifying the processing required at each step. 
[…]
