### 2024-10-02 #6775 @duxfrederic
Thanks for the thorough proofread @ivastar !

I implemented your requested changes. Regarding astropy: I am referring to a component in the base package at that line, so I think the citations in place are fine?

Ah, and I also skipped the "combining" suggestion, as the fluxes really are "combined" together? Basically a weighted mean of rescaled fluxes with sigma-clipping
