### 2024-10-02 #6775 @duxfrederic
Perfect! @ivastar

Here here the zenodo DOI: `10.5281/zenodo.13883045`, https://zenodo.org/records/13883045.

The post-review version is: v1.1.0, this is the archive that was uploaded to zenodo: https://github.com/duxfrederic/lightcurver/releases/tag/v1.1.0

The authors / ORCIDS / license all match.
