### 2024-10-08 #6775 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
