### 2024-10-08 #6775 @ivastar
Phew! And we are almost done! @duxfrederic thank you for your patience with this submission and sincere apologies that it took so long to finish the review.

Enormous thanks to @Onoddil and @benjaminrose for reviewing `lightcurver`!

The editor in chief will finalize the submission in the next few days.
