### 2024-10-12 82a2e82
Editorial tweaks to JOSS paper
