+++ paper.bib
@@ -16,18 +16,20 @@
 
 @ARTICLE{starredscience,
        author = {{Millon}, Martin and {Michalewicz}, Kevin and {Dux}, Fr{\'e}d{\'e}ric and {Courbin}, Fr{\'e}d{\'e}ric and {Marshall}, Philip J.},
-        title = "{Image deconvolution and PSF reconstruction with STARRED: a wavelet-based two-channel method optimized for light curve extraction}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
+        title = "{Image Deconvolution and Point-spread Function Reconstruction with STARRED: A Wavelet-based Two-channel Method Optimized for Light-curve Extraction}",
+      journal = {The Astronomical Journal},
+     keywords = {Deconvolution, Photometry, Light curves, Astronomy image processing, 1910, 1234, 918, 2306, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2024,
-        month = feb,
-          eid = {arXiv:2402.08725},
-        pages = {arXiv:2402.08725},
-          doi = {10.48550/arXiv.2402.08725},
+        month = aug,
+       volume = {168},
+       number = {2},
+          eid = {55},
+        pages = {55},
+          doi = {10.3847/1538-3881/ad4da7},
 archivePrefix = {arXiv},
        eprint = {2402.08725},
  primaryClass = {astro-ph.IM},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2024arXiv240208725M},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024AJ....168...55M},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
@@ -121,7 +123,7 @@ archivePrefix = {arXiv},
          {Weaver}, B.~A. and {Whitmore}, J.~B. and {Woillez}, J. and
          {Zabalza}, V. and {Astropy Contributors}},
         title = "{The Astropy Project: Building an Open-science Project and Status of the v2.0 Core Package}",
-      journal = {\aj},
+      journal = {The Astronomical Journal},
      keywords = {methods: data analysis, methods: miscellaneous, methods: statistical, reference systems, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2018,
         month = sep,
@@ -139,8 +141,8 @@ archivePrefix = {arXiv},
 
 @article{astropy_2022,
[…]
