### 2024-10-14 #6775 @duxfrederic
Thanks, @ivastar !

@warrickball Apologies, I had not seen your pull request immediately. Thanks for correcting the errors in the bibtex file.
I added these two lines before the final version; hopefully it is fine with you.
https://github.com/duxfrederic/lightcurver/commit/f880f18797486d2245367d43fa8ec7222cfbde2f
