### 2024-10-15 #6775 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
