### 2024-10-15 #6775 @warrickball
My apologies @duxfrederic but I missed that you've used monospace for `lightcurver` in the title of the article. Could you remove the backticks and let me know when you've done so? It leads to various metadata errors but I can rerun the acceptance command to fix things.
