### 2024-10-16 #6775 @warrickball
Yes... I was advised I could re-accept to fix the title. Perhaps I just need to do `accept` without trying `recommend-accept` first?
