### 2024-10-16 #6775 @xuanxu
the command to update a paper that is already accepted is `@editorialbot reaccept`, you can not recommend for acceptance a published submission. Sorry I didn't notice before.
