### 2024-10-16 #6775 @xuanxu
The command is `reaccept`, not `accept` 😄
