### 2024-10-16 #6775 @xuanxu
You have try it again, it only works if the command is in the first line of the comment, otherwise editorialbot ignores it.
