### 2024-10-16 #6775 @warrickball
Congratulations @duxfrederic, your paper has been (re)published in JOSS! (I apologise for my blundering about at the end that dragged it out.)
